--pg-listen-address and --pg-port should only be used for broadcasting an instance's address to the cluster state, but we should listen on all interfaces and port 5432 since we'll always be running within a Docker container.
Add IronNet Dockerfile and Makefile build scripts or creating the stolon
binaries in a container and storing it in ECR.

Make the default unix socket directory be /var/run/postgresql to match
the Debian version of PostgreSQL, which is what our container runs in.

Generate a pg_hba.conf that always trusts local connections for the
superuser for all database and replication connections.

Add new options to stolon-keeper to accept a directory for the
indexes and write-ahead log.